Prepare neurodamus for collective IO.
For neurodamus to use collective IO, all MPI ranks must read the same datasets,
the same number of times in the same order. This commit sorts all iterables of
attribute names to avoid issues.

The order is know to not be the same on all MPI ranks without this commit.